Moved samples/transfer_service and samples/transfer_job underneath projects, updated group samples transfer service to set error messages for unauthorized sample ids, updated tests
