+++ app/views/shared/samples/_errors.html.erb
@@ -1,5 +1,5 @@
 <div>
-  <%= viral_alert(type:, message:, classes: "mb-4 h-30 overflow-y-auto") do %>
+  <%= viral_alert(type:, message:, classes: "mb-4 max-h-50 overflow-y-auto") do %>
     <ul class="mt-1.5 ml-4 list-disc list-inside">
       <% errors.each do |error| %>
         <%- error_parts = error.split(":") %>
