+++ test/controllers/projects/samples/transfers_controller_test.rb
@@ -27,8 +27,8 @@ module Projects
         assert_response :unauthorized
       end
 
-      test 'should enqueue a Samples::TransferJob' do
-        assert_enqueued_jobs 1, only: ::Samples::TransferJob do
+      test 'should enqueue a Projects::Samples::TransferJob' do
+        assert_enqueued_jobs 1, only: Projects::Samples::TransferJob do
           post namespace_project_samples_transfer_path(@namespace, @project1, format: :turbo_stream),
                params: {
                  transfer: {
